+++ front/src/pages/Home.js
@@ -1,0 +1,251 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+import { withStyles, MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
+import Input from '@material-ui/core/Input';
+import InputBase from '@material-ui/core/InputBase';
+import InputAdornment from '@material-ui/core/InputAdornment';
+import InputLabel from '@material-ui/core/InputLabel';
+import IconButton from '@material-ui/core/IconButton';
+import Visibility from '@material-ui/icons/Visibility';
+import VisibilityOff from '@material-ui/icons/VisibilityOff';
+import FilledInput from '@material-ui/core/FilledInput';
+import MenuItem from '@material-ui/core/MenuItem';
+import TextField from '@material-ui/core/TextField';
+import Select from '@material-ui/core/Select';
+import FormControl from '@material-ui/core/FormControl';
+import purple from '@material-ui/core/colors/purple';
+import green from '@material-ui/core/colors/green';
+
+const styles = theme => ({
+  container: {
+    display: 'flex',
+    flexWrap: 'wrap',
+  },
+  root: {
+    display: 'flex',
+    flexWrap: 'wrap',
+  },
+  formControl: {
+    margin: theme.spacing.unit,
+    minWidth: 120,
+    maxWidth: 300,
+  },
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+    width: 200,
+  },
+  chips: {
+    display: 'flex',
+    flexWrap: 'wrap',
+  },
+  chip: {
+    margin: theme.spacing.unit / 4,
+  },
+  noLabel: {
+    marginTop: theme.spacing.unit * 3,
+  },
+  margin: {
+    margin: theme.spacing.unit,
+  },
+  cssLabel: {
+    '&$cssFocused': {
+      color: purple[500],
+    },
+  },
+  cssFocused: {},
+  cssUnderline: {
+    '&:after': {
+      borderBottomColor: purple[500],
+    },
+  },
+  cssOutlinedInput: {
+    '&$cssFocused $notchedOutline': {
+      borderColor: purple[500],
+    },
+  },
+  notchedOutline: {},
+  bootstrapRoot: {
+    'label + &': {
+      marginTop: theme.spacing.unit * 3,
+    },
+  },
+  bootstrapInput: {
+    borderRadius: 4,
+    backgroundColor: theme.palette.common.white,
+    border: '1px solid #ced4da',
+    fontSize: 16,
+    padding: '10px 12px',
+    transition: theme.transitions.create(['border-color', 'box-shadow']),
+    // Use the system font instead of the default Roboto font.
+    fontFamily: [
+      '-apple-system',
+      'BlinkMacSystemFont',
+      '"Segoe UI"',
+      'Roboto',
+      '"Helvetica Neue"',
+      'Arial',
+      'sans-serif',
+      '"Apple Color Emoji"',
+      '"Segoe UI Emoji"',
+      '"Segoe UI Symbol"',
+    ].join(','),
+    '&:focus': {
+      borderColor: '#80bdff',
+      boxShadow: '0 0 0 0.2rem rgba(0,123,255,.25)',
+    },
+  },
+  bootstrapFormLabel: {
+    fontSize: 18,
+  },
+});
+
+const theme = createMuiTheme({
+  palette: {
+    primary: green,
+  },
+  typography: { useNextVariants: true },
+});
+
+const classrooms = [101, 102, 103, 104, 105, 106, 201, 202, 203, 204];
+
+const ITEM_HEIGHT = 48;
+const ITEM_PADDING_TOP = 8;
+const MenuProps = {
+  PaperProps: {
+    style: {
+      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
+      width: 250,
+    },
+  },
+};
+
+function getStyles(classroom, that) {
+  return {
+    fontWeight:
+      that.state.classroom.indexOf(classroom) === -1
+        ? that.props.theme.typography.fontWeightRegular
+        : that.props.theme.typography.fontWeightMedium,
+  };
+}
+
+
+class Home extends React.Component{
+	state = {
+		classroom: [],
+    	showPassword: false,
+	}
+	handleChange = prop => event => {
+	    this.setState({ [prop]: event.target.value });
+	  };
+
+	handleClickShowPassword = () => {
+		this.setState(state => ({ showPassword: !state.showPassword }));
+	};
+
+	render(){
+		const { classes } = this.props;
+		return(
+			<div>
+		 		<header>
+		          <h1>Logo</h1>
+		        </header>
+		        <main className="home">
+		          <div className="courses">
+		            <h4>Courses</h4>
+		          </div>
+		        </main>
+		        <div className="modal">
+		          <TextField
+		            className={classes.margin}
+		            InputLabelProps={{
+		              classes: {
+		                root: classes.cssLabel,
+		                focused: classes.cssFocused,
+		              },
+		            }}
+		            InputProps={{
+		              classes: {
+		                root: classes.cssOutlinedInput,
+		                focused: classes.cssFocused,
+		                notchedOutline: classes.notchedOutline,
+		              },
+		            }}
+		            label="Nome do Curso"
+		            variant="outlined"
+		            id="course-name"
+		          />
+		          <TextField
+		            className={classes.margin}
+		            InputLabelProps={{
+		              classes: {
+		                root: classes.cssLabel,
+		                focused: classes.cssFocused,
+		              },
+		            }}
+		            InputProps={{
+		              classes: {
+		                root: classes.cssOutlinedInput,
+		                focused: classes.cssFocused,
+		                notchedOutline: classes.notchedOutline,
+		              },
+		            }}
+		            label="Professores"
+		            variant="outlined"
+		            id="teachers"
+		          />
+		          <FormControl variant="filled" className={classes.formControl}>
+		            <InputLabel
+		              ref={ref => {
+		                this.InputLabelRef = ref;
+		              }}
+		              htmlFor="classroom"
+		            >
+		              Salas
+		            </InputLabel>
+		            <Select
+		              multiple
+		              value={this.state.classroom}
+		              onChange={() => this.handleChange('classroom')}
+		              input={<FilledInput id="classroom" />}
+		              MenuProps={MenuProps}
+		            >
+		              {classrooms.map(classroom => (
+		                <MenuItem key={classroom} value={classroom} style={getStyles(classroom, this)}>
+		                  {classroom}
+		                </MenuItem>
+		              ))}
+		            </Select>
+		          </FormControl>
+		          <TextField
+		            id="start"
+		            label="Início"
+		            type="time"
+		            defaultValue="07:30"
+		            className={classes.textField}
+		            InputLabelProps={{
+		              shrink: true,
+		            }}
+		            inputProps={{
+		              step: 300, // 5 min
+		            }}
+		          />
+		          <TextField
+		            id="end"
+		            label="Fim"
+		            type="time"
+		            defaultValue="08:30"
+		            className={classes.textField}
+		            InputLabelProps={{
+		              shrink: true,
+		            }}
+		            inputProps={{
+		              step: 300, // 5 min
+		            }}
+		          />
+		        </div>
+	        </div>
+		)
+	}
+}
+export default withStyles(styles, { withTheme: true })(Home);
